pyproject.toml
[project]
name = "XAS-QQBot"
version = "2.1.1"
description = "Default template for PDM package"
readme = "README.md"
requires-python = "==3.12.*"
authors = [{ name = "XiYang6666", email = "1782356858@qq.com" }]
dependencies = [
    "nonebot-adapter-satori>=0.8.3",
    "nonebot2[fastapi,httpx,websockets]>=2.1.3",
    "filetype>=1.2.0",
    "typing-extensions>=4.9.0",
    "pillow>=10.2.0",
    "openai>=1.11.1",
    "nonebot-plugin-localstore>=0.5.2",
    "nonebot-plugin-apscheduler>=0.4.0",
    "mcstatus>=11.1.1",
    "httpx>=0.27.0",
]
license = { text = "MIT" }

[tool.nonebot]
adapters = [{ name = "Satori", module_name = "nonebot.adapters.satori" }]
plugins = ["nonebot_plugin_localstore", "nonebot_plugin_apscheduler"]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.pdm]
distribution = false

[tool.pylint]
disable = ["C0114", "C0115", "C0116", "C0301", "E0611"]

[tool.ruff]
line-length = 150
extend-select = ["I"]
